chore: Fix VS compilation warnings

+++ src/game/gui/debug.cpp
@@ -54,10 +54,10 @@ void DebugGui::update(const DebugContext &ctx) {
 
         float textWidth = _font->measure(object.text);
         Control::Extent extent(
-            _gfxOpts.width * object.screenCoords.x - 0.5f * textWidth,
-            _gfxOpts.height * (1.0f - object.screenCoords.y),
-            textWidth,
-            _font->height());
+            static_cast<int>(_gfxOpts.width * object.screenCoords.x - 0.5f * textWidth),
+            static_cast<int>(_gfxOpts.height * (1.0f - object.screenCoords.y)),
+            static_cast<int>(textWidth),
+            static_cast<int>(_font->height()));
 
         Control::Text text;
         text.text = object.text;
